requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
kaggle>=1.6.0